+++ common/src/main/java/smartin/miapi/client/gui/crafting/CraftingScreen.java
@@ -1,6 +1,7 @@
 package smartin.miapi.client.gui.crafting;
 
 import com.mojang.blaze3d.systems.RenderSystem;
+import net.bettercombat.logic.WeaponAttributesFallback;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.screen.ingame.ScreenHandlerProvider;
@@ -14,12 +15,16 @@ import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.Nullable;
 import smartin.miapi.Miapi;
 import smartin.miapi.blocks.ModularWorkBenchEntity;
-import smartin.miapi.client.gui.*;
+import smartin.miapi.client.gui.InteractAbleWidget;
+import smartin.miapi.client.gui.ParentHandledScreen;
+import smartin.miapi.client.gui.SimpleScreenHandlerListener;
+import smartin.miapi.client.gui.TransformableWidget;
 import smartin.miapi.client.gui.crafting.crafter.DetailView;
 import smartin.miapi.client.gui.crafting.crafter.ModuleCrafter;
 import smartin.miapi.client.gui.crafting.slotdisplay.SlotDisplay;
 import smartin.miapi.client.gui.crafting.slotdisplay.SmithDisplay;
-import smartin.miapi.client.gui.crafting.statdisplay.StatListWidget;
+import smartin.miapi.client.gui.crafting.statdisplay.material.MaterialStatWidget;
+import smartin.miapi.client.gui.crafting.statdisplay.material.StatDisplayWidget;
 import smartin.miapi.item.ModularItemStackConverter;
 import smartin.miapi.modules.ItemModule;
 import smartin.miapi.modules.ModuleInstance;
@@ -29,6 +34,7 @@ import smartin.miapi.modules.properties.AllowedSlots;
 import smartin.miapi.modules.properties.SlotProperty;
 import smartin.miapi.registries.RegistryInventory;
 
+import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -38,7 +44,8 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
     public static final InterpolateMode EASE_OUT = new InterpolateMode.EaseOut(5);*/
     private ItemStack stack;
     private ModuleCrafter moduleCrafter;
-    private StatListWidget statDisplay;
+    private StatDisplayWidget statDisplay;
+    private MaterialStatWidget materialStatWidget;
     private SlotDisplay slotDisplay;
     private SmithDisplay smithDisplay;
     private MinimizeButton minimizer;
@@ -46,20 +53,33 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
     @Nullable
     public static SlotProperty.ModuleSlot slot;
     @Nullable
-    private static EditOption editOption;
+    public static EditOption editOption;
     private TransformableWidget editHolder;
     static int editSpace = 30;
+    @Nullable
+    public InteractAbleWidget hoverElement = null;
+    static WeakReference<CraftingScreen> craftingScreenWeakReference = new WeakReference<>(null);
 
     List<InteractAbleWidget> editOptionIcons = new ArrayList<>();
 
     public CraftingScreen(CraftingScreenHandler handler, PlayerInventory inventory, Text title) {
         super(handler, inventory, Text.empty());
         slot = null;
-        this.backgroundWidth = 369 + 12 - 15;
+        this.backgroundWidth = 369 + 12 - 15 + 6;
 
         this.backgroundHeight = 223 - 9 - 15;
 
         DetailView.scrollPos = 0;
+        craftingScreenWeakReference = new WeakReference<>(this);
+    }
+
+    @Nullable
+    public static CraftingScreen getInstance() {
+        if (craftingScreenWeakReference != null && craftingScreenWeakReference.get() != null) {
+            return craftingScreenWeakReference.get();
+        }
+        WeaponAttributesFallback fallback;
+        return null;
     }
 
     public EditOption getEditOption() {
@@ -91,7 +111,7 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
         allowedModules.add("melee");
         baseSlot = new SlotProperty.ModuleSlot(allowedModules);
 
-        int centerX = (this.width - this.backgroundWidth) / 2;
+        int centerX = (this.width - this.backgroundWidth - 6) / 2;
         int centerY = (this.height - this.backgroundHeight) / 2;
 
         moduleCrafter = new ModuleCrafter(centerX + 51 - 15, centerY + 22 - 14, 144, 89, this::selectSlot, (item) -> {
@@ -103,14 +123,14 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
         moduleCrafter.setPacketIdentifier(handler.packetID);
         this.addChild(moduleCrafter);
 
-        slotDisplay = new SlotDisplay(stack, centerX + 51 - 15, centerY + 117 - 14, 68, 87, (selected) -> {
+        slotDisplay = new SlotDisplay(ItemStack.EMPTY, centerX + 51 - 15, centerY + 117 - 14, 68, 87, (selected) -> {
 
         });
         slotDisplay.setItem(getItem());
         this.addChild(slotDisplay);
         smithDisplay = new SmithDisplay(centerX + 140 - 15, centerY + 117 - 14, 55, 70);
         this.addChild(smithDisplay);
-        statDisplay = new StatListWidget(centerX + 213 - 15, centerY + 30 - 14, 161, 95);
+        statDisplay = new StatDisplayWidget(centerX + 213 - 15, centerY + 30 - 14, 161, 95);
         this.addChild(statDisplay);
 
         minimizer = new MinimizeButton(centerX + 178 - 15, centerY + 188 - 14, 18, 18, this::minimizeView, this::maximizeView);
@@ -139,6 +159,14 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
         previewStack(handler.inventory.getStack(0));
         PreviewManager.resetCursorStack();
         PreviewManager.resetPreview();
+    }
+
+    public int getBackgroundHeight() {
+        return this.backgroundHeight;
+    }
+
+    public int getBackgroundWidth() {
+        return this.backgroundWidth;
     }
 
     //could be the same as maximizeView()
@@ -192,8 +220,12 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
         handler.inventory.setStack(0, stack);
     }
 
-    private void updateItem(ItemStack stack) {
+    public void updateItem(ItemStack stack) {
         PreviewManager.resetCursorStack();
+        updatePreviewItemStack(stack);
+    }
+
+    public void updatePreviewItemStack(ItemStack stack) {
         stack = stack.copy();
         slotDisplay.setItem(stack);
         ItemStack converted = ModularItemStackConverter.getModularVersion(stack).copy();
@@ -220,8 +252,7 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
             smithDisplay.setPreview(converted);
         }
         if (statDisplay != null) {
-            statDisplay.setOriginal(converted);
-            statDisplay.setCompareTo(converted);
+            statDisplay.setItemsOriginal(converted, converted);
         }
         /*
         List<Integer> slotPos = new ArrayList<>();
@@ -336,7 +367,7 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
     public void render(DrawContext drawContext, int mouseX, int mouseY, float delta) {
         RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
         renderBackground(drawContext);
-        int i = (this.width - this.backgroundWidth) / 2;
+        int i = (this.width - this.backgroundWidth - 6) / 2;
         int j = (this.height - this.backgroundHeight) / 2;
         //InteractAbleWidget.drawSquareBorder(drawContext, i, j, this.backgroundWidth, this.backgroundHeight, 1, ColorHelper.Argb.getArgb(255, 255, 0, 0));
         //(Identifier texture, int x, int y, int width, int height, float u, float v, int regionWidth, int regionHeight, int textureWidth, int textureHeight)
@@ -367,15 +398,66 @@ public class CraftingScreen extends ParentHandledScreen<CraftingScreenHandler> i
 
             drawContext.drawTexture(BACKGROUND_TEXTURE, i + 43 - 15, j + 111 - 14, 160, 95, 0, 199, 160, 95, 512, 512);
         }
-        super.render(drawContext, mouseX, mouseY, delta);
+        if (hoverElement == null) {
+            super.render(drawContext, mouseX, mouseY, delta);
+        } else {
+            hoverElement.render(drawContext, mouseX, mouseY, delta);
+        }
         this.drawMouseoverTooltip(drawContext, mouseX, mouseY);
         drawContext.getMatrices().push();
         drawContext.getMatrices().translate(0.0F, 0.0F, 400.0F);
-        drawContext.draw(() -> {
-            this.renderHover(drawContext, mouseX, mouseY, delta);
-        });
+        if (hoverElement == null) {
+            drawContext.draw(() -> {
+                this.renderHover(drawContext, mouseX, mouseY, delta);
+            });
+        } else {
+            drawContext.draw(() -> {
+                hoverElement.renderHover(drawContext, mouseX, mouseY, delta);
+            });
+        }
         drawContext.getMatrices().pop();
         PreviewManager.tick();
+    }
+
+    @Override
+    public boolean mouseClicked(double mouseX, double mouseY, int button) {
+        if (hoverElement != null) {
+            return hoverElement.mouseClicked(mouseX, mouseY, button);
+        }
+        return super.mouseClicked(mouseX, mouseY, button);
+    }
+
+    @Override
+    public void mouseMoved(double mouseX, double mouseY) {
+        if (hoverElement != null) {
+            hoverElement.mouseMoved(mouseX, mouseY);
+        } else {
+            super.mouseMoved(mouseX, mouseY);
+        }
+    }
+
+    @Override
+    public boolean mouseReleased(double mouseX, double mouseY, int button) {
+        if (hoverElement != null) {
+            return hoverElement.mouseReleased(mouseX, mouseY, button);
+        }
+        return super.mouseReleased(mouseX, mouseY, button);
+    }
+
+    @Override
+    public boolean mouseDragged(double mouseX, double mouseY, int button, double deltaX, double deltaY) {
+        if (hoverElement != null) {
+            return hoverElement.mouseDragged(mouseX, mouseY, button, deltaX, deltaY);
+        }
+        return super.mouseDragged(mouseX, mouseY, button, deltaX, deltaY);
+    }
+
+    @Override
+    public boolean mouseScrolled(double mouseX, double mouseY, double amount) {
+        if (hoverElement != null) {
+            return hoverElement.mouseScrolled(mouseX, mouseY, amount);
+        }
+        return super.mouseScrolled(mouseX, mouseY, amount);
     }
 
     @Override