+++ common/src/main/java/smartin/miapi/client/gui/crafting/statdisplay/StatListWidget.java
@@ -6,7 +6,6 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.Element;
-import net.minecraft.client.gui.widget.ClickableWidget;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
@@ -373,7 +372,7 @@ public class StatListWidget extends InteractAbleWidget {
     }
 
     public StatListWidget(int x, int y, int width, int height) {
-        super(x, y, width, height, Text.empty());
+        super(x, y, width, height, Text.literal("miapi.module.statdisplay"));
         transformableWidget = new TransformableWidget(x, y, width, height, Text.empty());
         boxList = new BoxList(x, y, width, height, Text.empty(), new ArrayList<>());
         ScrollList list = new ScrollList(x, y, width, height, List.of(boxList));
@@ -414,13 +413,21 @@ public class StatListWidget extends InteractAbleWidget {
     }
 
     private <T extends InteractAbleWidget & SingleStatDisplay> void update() {
+        setAttributeCaches(original, compareTo);
+        boxList.setWidgets(collectWidgets(original, compareTo), 0);
+    }
+
+    public static void setAttributeCaches(ItemStack original, ItemStack compareTo) {
         for (EquipmentSlot equipmentSlot : EquipmentSlot.values()) {
             Multimap<EntityAttribute, EntityAttributeModifier> oldAttr = original.getAttributeModifiers(equipmentSlot);
             Multimap<EntityAttribute, EntityAttributeModifier> compAttr = compareTo.getAttributeModifiers(equipmentSlot);
             AttributeSingleDisplay.oldItemCache.put(equipmentSlot, oldAttr);
             AttributeSingleDisplay.compareItemCache.put(equipmentSlot, compAttr);
         }
-        List<ClickableWidget> widgets = new ArrayList<>();
+    }
+
+    public static <T extends InteractAbleWidget & SingleStatDisplay> List<InteractAbleWidget> collectWidgets(ItemStack original, ItemStack compareTo) {
+        List<InteractAbleWidget> widgets = new ArrayList<>();
         for (StatWidgetSupplier supplier : statWidgetSupplier) {
             List<T> statWidgets = supplier.currentList(original, compareTo);
             for (T statDisplay : statWidgets) {
@@ -439,7 +446,26 @@ public class StatListWidget extends InteractAbleWidget {
 
             }
         }
-        boxList.setWidgets(widgets, 0);
+        List<InteractAbleWidget> sortedWidgets = new ArrayList<>();
+        int shortSize = 80;
+        InteractAbleWidget buffered = null;
+        for (InteractAbleWidget widget : widgets) {
+            if (widget.getWidth() <= shortSize) {
+                if (buffered == null) {
+                    buffered = widget;
+                } else {
+                    sortedWidgets.add(buffered);
+                    sortedWidgets.add(widget);
+                    buffered = null;
+                }
+            } else {
+                sortedWidgets.add(widget);
+            }
+        }
+        if (buffered != null) {
+            sortedWidgets.add(buffered);
+        }
+        return sortedWidgets;
     }
 
     public interface StatWidgetSupplier {
@@ -464,6 +490,12 @@ public class StatListWidget extends InteractAbleWidget {
     }
 
     public void setCompareTo(ItemStack compareTo) {
+        this.compareTo = compareTo;
+        update();
+    }
+
+    public void setItemsOriginal(ItemStack original, ItemStack compareTo) {
+        this.original = original;
         this.compareTo = compareTo;
         update();
     }