+++ common/src/main/java/smartin/miapi/modules/material/Material.java
@@ -68,9 +68,32 @@ public interface Material {
         return false;
     }
 
+    /**
+     * Retuns all Material Properties for this key, see {@link MaterialProperties} for more details
+     * @param key
+     * @return
+     */
     Map<ModuleProperty, JsonElement> materialProperties(String key);
 
+    /**
+     * be sure to also implement {@link Material#getAllDisplayPropertyKeys()}
+     * if you implement this
+     * @param key
+     * @return
+     */
+    default Map<ModuleProperty, JsonElement> getDisplayMaterialProperties(String key) {
+        return materialProperties(key);
+    }
+
+    /**
+     * retuns all unique property keys this Material has Properties for.
+     * @return
+     */
     List<String> getAllPropertyKeys();
+
+    default List<String> getAllDisplayPropertyKeys() {
+        return getAllPropertyKeys();
+    }
 
     double getDouble(String property);